Stop using shorts and chars where ints are better
C implicitly converts variables to int all the time, so using chars
and shorts everywhere will just result in a ton of unnecessary
casting.

This conversion only extends to local variables and function
parameters: state variables still use the smallest available types,
as state is optimized for size rather than performance.

This eliminates all warnings generated when compiling clownmdemu with
MSVC 12.0/Visual Studio 6.

Sidenote: it's so cool to finally be able to have MSVC 12.0 compile
my code as C for once. Usually, I program in a common subset of C99
and C++98, and rely on compiling my code as C++ in order for building
with MSVC 12.0 to work. With C89, it can just compile as C instead.