Make some local const arrays static
Some compilers (such as GCC) are stupid and force local const arrays
to be copied to the stack, even though they're available in global
memory for literally the entire lifetime of the program.

It's best to prevent these stupid compilers from being stupid by
forcing them to avoid this behaviour with the 'static' keyword.

In the ideal case that the compiler was already reading these arrays
directly from global memory, then this altered code will function
identically.