Un-generate condition code logic.
This is slower than the previous code (even moreso if the lookup
table is not used), but once the refactor is complete, the
interpreter should be faster overall.

This is the first part of a big new refactor: rather than generating
an 'unrolled' interpreter, I want to compute a gigantic lookup table
of function pointers, with each function pointer representing a
'step' of instruction execution. This is a lot like the original
version of the interpreter, which used giant switch statements in
this way, but this new method aims to improve upon it by eliminating
the overhead of using a series of switch statements, and instead use
a single giant lookup table.