Turn supervisor check into a function.
This is the first of many commits that aim to reduce each
instruction's unique logic to a series of function calls, with the
eventual goal of replacing the generated code with a series of
function pointer lists. This will be very important for reengineering
this emulator to be oriented around bus events rather than
instructions.